\name{AUC}
\alias{AUC}
\title{Compute AUC}
\usage{
  AUC(x, y, index = NULL)
}
\arguments{
  \item{x}{Numeric vector.}

  \item{y}{Numeric vector.}

  \item{index}{possible index vector if only a subset
  should be used.}
}
\description{
  Compute AUC (area under the curve) using the trapezoidal
  rule.
}
\author{
  Benjamin Hofner
}
\keyword{internal}

