\name{default_color_regions}
\alias{default_color_regions}
\title{Colour regions}
\usage{
  default_color_regions(colors = NULL)
}
\arguments{
  \item{colors}{Character or integer vector with at least
  two distinct colours. If \code{NULL} or empty, default
  colours are chosen.}
}
\value{
  Character vector of colour codes.
}
\description{
  Create default colour regions for use with
  \code{\link{level_plot}}.
}
\keyword{internal}

