\name{extract_curve_params.grofit}
\alias{extract_curve_params.grofit}
\title{Grofit extraction}
\usage{
  \method{extract_curve_params}{grofit} (data)
}
\arguments{
  \item{data}{Object of class \sQuote{grofit}.}
}
\value{
  See \code{\link{curve_params}} (in \code{full} mode).
}
\description{
  Extract and rename estimated curve parameters.
}
\keyword{internal}

