\name{fit_spline}
\alias{fit_spline}
\title{Fit splines}
\usage{
  fit_spline(y, x = "Hour", data,
    options = set_spline_options(), weights = NULL)
}
\arguments{
  \item{y}{Character scalar. Outcome. Usually the name of a
  well.}

  \item{x}{Character scalar. Predictor. Usually
  \code{"Hour"}.}

  \item{data}{A \code{data.frame} containing the variables
  specified in \code{x} and \code{y}.}

  \item{options}{List of options. Generated by
  \code{\link{set_spline_options}}.}

  \item{weights}{vector of weights used in the fitting
  process.}
}
\value{
  Spline fit. A model of class \code{"opm_model"}.
}
\description{
  Fit various splines (smoothing splines and P-splines from
  \pkg{mgcv} and smoothing splines via
  \code{smooth.spline}) to opm data.
}
\author{
  Benjamin Hofner
}
\seealso{
  \code{\link{gam}} from \pkg{mgcv} and
  \code{\link{smooth.spline}}
}
\keyword{internal}

