\docType{class}
\name{FOE}
\alias{FOE}
\alias{FOE-class}
\title{FOE class}
\description{
  This is a virtual class facilitating the implementation
  of functionality for both formulas and expressions.
  Methods defined for objects from the class can be applied
  to either kind of object, but this class is not directly
  dealt with by an \pkg{opm} user.
}
\details{
  See \code{\link{metadata.set}} and
  \code{\link{map_metadata}} for usage examples.
}
\note{
  \sQuote{FOE} is an acronym for \sQuote{formula or
  expression}.
}
\seealso{
  methods::Methods base::expression stats::formula

  Other classes: \code{\link{MOA}}, \code{\link{OPM}},
  \code{\link{OPMA}}, \code{\link{OPMD}},
  \code{\link{OPMS}}, \code{\link{OPMX}},
  \code{\link{WMD}}, \code{\link{YAML_VIA_LIST}}
}
\keyword{classes}
\keyword{methods}

