\name{get_data}
\alias{get_data}
\alias{get_data.psp_model}
\alias{get_data.smooth.spline_model}
\title{Extract data}
\usage{
  get_data(x)

  \method{get_data}{psp_model} (x)

  \method{get_data}{smooth.spline_model} (x)
}
\arguments{
  \item{x}{Model.}
}
\value{
  data.frame
}
\description{
  Extract data from various spline estimates
}
\author{
  Benjamin Hofner
}
\keyword{internal}

