\name{indexes}
\alias{indexes}
\title{Collect factor indexes}
\arguments{
  \item{x}{Factor.}
}
\value{
  List of integer vectors
}
\description{
  Convert a factor to a list with its levels as keys and
  the positions of each level within the factor as values.
}
\keyword{internal}

\docType{methods}
\alias{indexes-methods}
\alias{indexes,factor-method}
\usage{
  \S4method{indexes}{factor}(x) 

}
