\name{initialize,OPMS}
\alias{initialize,OPMS}
\title{Initialize}
\arguments{
  \item{.Object}{\code{\link{OPMS}} object.}

  \item{...}{Additional arguments.}
}
\value{
  \code{\link{OPMS}} object.
}
\description{
  Initialize method for the \code{\link{OPMS}} class.
}
\keyword{internal}

\docType{methods}
\alias{initialize,OPMS-method}
\usage{
  \S4method{initialize}{OPMS}(.Object, ...) 
}
