\name{initialize}
\alias{initialize}
\title{Initialize}
\arguments{
  \item{.Object}{\code{\link{OPM}} or \code{\link{OPMS}}
  object.}

  \item{...}{Additional arguments.}
}
\value{
  \code{\link{OPM}} or \code{\link{OPMS}} object.
}
\description{
  Initialize methods for the \code{\link{OPM}} and
  \code{\link{OPMS}} classes.
}
\keyword{internal}

\docType{methods}
\alias{initialize-methods}
\alias{initialize,OPM-method}
\alias{initialize,OPMS-method}
\usage{
  \S4method{initialize}{OPM}(.Object, ...) 

  \S4method{initialize}{OPMS}(.Object, ...) 

}
