\name{is_constant}
\alias{is_constant}
\title{Check for constantness.}
\usage{
  is_constant(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An R object to which \code{duplicated} can be
  applied.}

  \item{na.rm}{Logical scalar. Remove \code{NA} elements
  before determining constantness?}
}
\value{
  Logical scalar.
}
\description{
  Assess whether all elements in a collection are
  identical.
}
\keyword{internal}

