\name{last}
\alias{last}
\title{Fetch the last elements}
\usage{
  last(x, i = 1L)
}
\arguments{
  \item{x}{An \R object to which \code{[} can be applied.}

  \item{i}{Integer scalar. Number of elements to fetch.}
}
\value{
  Object of the same class than \code{x}.
}
\description{
  Fetch the last element(s) from a subsettable object.
}
\keyword{internal}

