\name{main_title}
\alias{main_title}
\title{Create main title}
\arguments{
  \item{object}{\code{\link{OPMX}} object.}

  \item{settings}{See the \code{main} argument of
  \code{\link{xy_plot}}.}
}
\value{
  Character scalar or \code{NULL}.
}
\description{
  Create a title used as value of the \sQuote{main}
  argument of the plotting functions.
}
\keyword{internal}

\docType{methods}
\alias{main_title-methods}
\alias{main_title,OPMX-method}
\usage{
  \S4method{main_title}{OPMX}(object, settings) 

}
