\name{md_data_frame}
\alias{md_data_frame}
\title{Convert to metadata-like data frame}
\usage{
  md_data_frame(object, stringsAsFactors, optional, ...)
}
\arguments{
  \item{object}{Nested list.}

  \item{stringsAsFactors}{Logical scalar.}

  \item{optonal}{Logical scalar.}
}
\value{
  Data frame.
}
\description{
  A helper function for \code{\link{to_metadata}}.
}
\keyword{internal}

