\docType{class}
\name{MOA}
\alias{MOA}
\alias{MOA-class}
\title{MOA class}
\description{
  This is a virtual class facilitating the implementation
  of functionality for both matrices and arrays. Methods
  defined for objects from the class can be applied to
  either kind of object, but this class is not directly
  dealt with by an \pkg{opm} user.
}
\details{
  \sQuote{MOA} is an acronym for \sQuote{matrix or array}.
}
\seealso{
  methods::Methods base::matrix base::array

  Other classes: \code{\link{OPM}}, \code{\link{OPMA}},
  \code{\link{OPMD}}, \code{\link{OPMS}},
  \code{\link{OPMX}}, \code{\link{WMD}},
  \code{\link{YAML_VIA_LIST}}
}
\keyword{classes}
\keyword{methods}

