\name{other_slots}
\alias{other_slots}
\title{Other slots}
\arguments{
  \item{object}{\code{\link{OPM}} object.}
}
\value{
  Character vector.
}
\description{
  Return the names of all slots except the one holding the
  measurements.
}
\keyword{internal}

\docType{methods}
\alias{other_slots-methods}
\alias{other_slots,OPM-method}
\usage{
  \S4method{other_slots}{OPM}(object) 

}
