\name{predict.smooth.spline_model}
\alias{predict.smooth.spline_model}
\title{Predictions for smoothing spline models}
\usage{
  \method{predict}{smooth.spline_model} (object,
    newdata = NULL, ...)
}
\arguments{
  \item{object}{An object of class
  \code{smooth.spline_model}.}

  \item{newdata}{data.frame containing new data for
  predictions. This is optional.}

  \item{...}{additional options. Currently not used.}
}
\description{
  Make prediction for estimated splines.
}
\author{
  Benjamin Hofner
}
\keyword{internal}

