\name{read_old_opm}
\alias{read_old_opm}
\title{Read old-style OmniLog CSV file}
\usage{
  read_old_opm(filename)
}
\arguments{
  \item{filename}{Character scalar with the obvious
  meaning.}
}
\value{
  \code{\link{OPM}} object.
}
\description{
  Read OmniLog\eqn{\textsuperscript{\textregistered}}{(R)}
  data from single pseudo-\acronym{CSV} file in the old
  (pre-2011) format.
}
\references{
  \url{http://www.biolog.com/}
}
\keyword{internal}

