\name{repair_na_strings,character}
\alias{repair_na_strings,character}
\title{Repair NAs}
\arguments{
  \item{object}{Character vector.}
}
\value{
  Character vector.
}
\description{
  Replace \sQuote{NA} by \code{NA_character_}.
}
\keyword{internal}

\docType{methods}
\alias{repair_na_strings,character-method}
\usage{
  \S4method{repair_na_strings}{character}(object) 
}
