\name{smooth.construct.psp.smooth.spec}
\alias{smooth.construct.psp.smooth.spec}
\title{Constructor for P-splines}
\usage{
  \method{smooth.construct}{psp.smooth.spec} (object, data,
    knots)
}
\arguments{
  \item{object}{a smooth specification object, usually
  generated by the term \code{s(x, bs = "psp", ...)}.}

  \item{data}{a list containing just the data required by
  this term with names corresponding to
  \code{object$term}.}

  \item{knots}{a list containing any knots supplied for
  basis setup in same order and with same names as
  \code{data}.}
}
\value{
  An object of class \code{psp.smooth}.
}
\description{
  A modified version of the P-spline constructor
  \code{\link{smooth.construct.ps.smooth.spec}} from
  package \pkg{mgcv} that allows to make at or outside of
  the boundary knots.
}
\details{
  This function is primarily for internal use. For details
  see \code{\link{smooth.construct}} and
  \code{\link{smooth.construct.ps.smooth.spec}}.
}
\author{
  Benjamin Hofner bases on work by Simon N. Wood
}
\keyword{internal}

