\name{test_file_dir}
\alias{test_file_dir}
\title{Test file directory}
\usage{
  test_file_dir(x = source_location(), files = NULL)
}
\arguments{
  \item{x}{Location of source file.}

  \item{files}{Character vector. Optional list of of
  filenames to append to the directory name.}
}
\value{
  Name of the directory in which the input files for
  testing reside.
}
\description{
  Name of directory with test files. Auxiliary function.
}
\keyword{internal}

