\name{to_sentence}
\alias{to_sentence}
\alias{to_sentence.logical}
\title{Create sentences}
\usage{
  to_sentence(x, ...)

  \method{to_sentence}{logical} (x, html, ...)
}
\arguments{
  \item{x}{Logical vector.}

  \item{html}{Logical scalar.}

  \item{...}{Optional arguments passed to and from other
  methods.}
}
\value{
  Character vector, one element per sentence.
}
\description{
  Create a textual (listing-like) description.
}
\keyword{internal}

