\name{update_settings_list}
\alias{update_settings_list}
\title{Update settings entries}
\arguments{
  \item{x}{List.}
}
\value{
  List.
}
\description{
  This converts old-style to new-style aggregation or
  discretization settings.
}
\keyword{internal}

\docType{methods}
\alias{update_settings_list,list-method}
\usage{
  \S4method{update_settings_list}{list}(x) 

}
