\docType{class}
\name{YAML_VIA_LIST}
\alias{YAML_VIA_LIST}
\alias{YAML_VIA_LIST-class}
\title{YAML_VIA_LIST class}
\description{
  This is a virtual class facilitating the conversion to
  \acronym{YAML} format (or its subset, \acronym{JSON}). It
  can currently be used by any class that can be coerced to
  a list, but it is not directly applied by an \pkg{opm}
  user.
}
\details{
  See \code{\link{to_yaml}} for a usage example.
}
\seealso{
  methods::Methods

  Other classes: \code{\link{FOE}}, \code{\link{MOA}},
  \code{\link{OPM}}, \code{\link{OPMA}},
  \code{\link{OPMD}}, \code{\link{OPMS}},
  \code{\link{OPMX}}, \code{\link{WMD}}
}
\keyword{classes}
\keyword{methods}

