\name{default_color_regions}
\alias{default_color_regions}
\title{Color regions}
\usage{
  default_color_regions(colors = NULL)
}
\arguments{
  \item{colors}{Character or integer vector with at least
  two distinct colors.  If \code{NULL} or empty, default
  colors are chosen.}
}
\value{
  Character vector of color codes.
}
\description{
  Create default color regions for use with
  \code{\link{level_plot}}.
}
\keyword{internal}

