\name{extract_curve_params}
\alias{extract_curve_params}
\alias{extract_curve_params.grofit}
\title{Grofit extraction}
\usage{
  extract_curve_params(x)

  \method{extract_curve_params}{grofit} (x)
}
\arguments{
  \item{x}{Object of class \sQuote{grofit}.}
}
\value{
  Matrix.
}
\description{
  Extract and rename estimated curve parameters.
}
\keyword{internal}

