\name{initialize}
\alias{initialize}
\title{Initialize}
\arguments{
  \item{.Object}{\code{\link{OPM}} object.}

  \item{...}{Additional arguments.}
}
\value{
  \code{\link{OPM}} object.
}
\description{
  Initialize method for the \code{\link{OPM}} class.
}
\keyword{internal}

\docType{methods}
\alias{initialize-methods}
\alias{initialize,OPM-method}
\usage{
  \S4method{initialize}{OPM}(.Object, ...) 

}
