\docType{class}
\name{OPMX}
\alias{OPMX}
\alias{OPMX-class}
\title{OPMX class}
\description{
  This is a virtual class containing helper methods mainly
  for plotting \code{\link{OPM}} and \code{\link{OPMS}}
  objects. It is not directly applied by an \pkg{opm} user.
}
\details{
  See \code{\link{show}} and \code{\link{sort}} for usage
  examples.
}
\note{
  Regarding the name: \acronym{OPMX} stands for
  \sQuote{\acronym{OPM} or \acronym{OPMS}}.
}
\seealso{
  methods::Methods

  Other classes: \code{\link{FOE}}, \code{\link{MOA}},
  \code{\link{OPM}}, \code{\link{OPMA}},
  \code{\link{OPMD}}, \code{\link{OPMS}},
  \code{\link{WMD}}, \code{\link{YAML_VIA_LIST}}
}
\keyword{classes}
\keyword{methods}

