\name{param_names}
\alias{param_names}
\title{Names of curve parameters}
\usage{
  param_names()
}
\value{
  Character vector.
}
\description{
  Yield the names of the estimated curve parameters used
  internally and in the output.
}
\examples{
(x <- param_names())
stopifnot(is.character(x), length(x) == 4L, identical(unique(x), x))
}
\seealso{
  Other aggregation-functions: \code{\link{do_aggr}},
  \code{\link{fast_estimate}}
}
\keyword{utilities}

