\name{print}
\alias{print}
\alias{print.OPMD_Listing}
\alias{print.OPMS_Listing}
\alias{print.OPMS_Summary}
\alias{print.OPM_Summary}
\title{Print}
\usage{
  \method{print}{OPMD_Listing} (x, ...)

  \method{print}{OPMS_Listing} (x, ...)

  \method{print}{OPM_Summary} (x, ...)

  \method{print}{OPMS_Summary} (x, ...)
}
\arguments{
  \item{x}{Object of class \sQuote{OPM_Summary} or
  \sQuote{OPMS_Summary}.}

  \item{...}{Optional arguments passed to and from other
  methods.}
}
\value{
  \code{x} is returned invisibly.
}
\description{
  Print \code{\link{OPM}} or \code{\link{OPMS}} summaries
  to the screen.
}
\keyword{internal}

