\name{read_new_opm}
\alias{read_new_opm}
\title{Read OmniLog CSV file}
\usage{
  read_new_opm(filename)
}
\arguments{
  \item{filename}{Character scalar with the obvious
  meaning.}
}
\value{
  \code{\link{OPM}} object.
}
\description{
  Read OmniLog\eqn{\textsuperscript{\textregistered}}{(R)}
  data from single \acronym{CSV} file in novel (2011)
  format.
}
\references{
  \url{http://www.biolog.com/}
}
\keyword{internal}

