\name{source_location}
\alias{source_location}
\title{Source file name}
\usage{
  source_location()
}
\value{
  Name of file in which this function is defined or
  \code{NULL}.
}
\description{
  Name of this source file. Does not work if \code{library}
  instead of \code{source} is used. Auxiliary function.
}
\keyword{internal}

