\name{test_file_dir}
\alias{test_file_dir}
\title{Test file directory}
\usage{
  test_file_dir(loc)
}
\arguments{
  \item{loc}{Location of source file.}
}
\value{
  Name of the directory in which the input files for
  testing reside.
}
\description{
  Name of directory with test files. Auxiliary function.
}
\seealso{
  Other testing-functions: \code{\link{source_location}}
}
\keyword{internal}

