\name{to_sentence}
\alias{to_sentence}
\alias{to_sentence.logical}
\title{Create sentences}
\usage{
  to_sentence(x)

  \method{to_sentence}{logical} (x)
}
\arguments{
  \item{x}{Logical vector.}
}
\value{
  Character vector, one element per sentence.
}
\description{
  Create a textual (listing-like) description.
}
\keyword{internal}

